nixos/lib/utils: add /home to dirs required for boot
The user home directories for users specified in `users.users` are
created by an activation script in the initrd, this requires that
`/home` is mounted properly. Otherwise `/home` and the user directories
in it might be created on the wrong filesystem with the right filesystem
being mounted over it after the initrd. This can lead to missing user
home directories.